[BaaS Errors] Fixed BaaSExceptionFactory Bug that Caused Services to Crash
App: baas-errors, baas-account-service, baas-customer-service, core-bank-simulator

Fixed a bug that caused all of the BaaS services to crash when the
service received an AxiosError. The static BaaSExceptionFactory.create
method was not able to identify the BaaSErrors when running
because it is static. Moved the logic to "createBaaSException" function
and replaced all occurrences of "BaaSExceptionFactory.create"
with "createBaaSExption" calls.