[Account] Verify Customer is Active before Creating Account
Apps: baas-account-service, baas-errors, core-bank-simulator

Before creating an account verify that all of the participants are active
customers by calling the GET /v1/customers/:customerId for each
participant_customer_id in the create account request. If all of the
participant_customer_ids are not active respnd with a Bad Request error.
After validating the customer_ids then fill in the name_on_account and
account_name fields in the create account DTO.
=>
On branch validate-account-customers
Changes to be committed:
	modified:   apps/baas-account-service/src/accounts/accounts.module.ts
	modified:   apps/baas-account-service/src/accounts/accounts.service.ts
	modified:   apps/baas-account-service/src/accounts/dto/create-account.dto.ts
	modified:   apps/baas-account-service/src/config/configuration.ts
	new file:   apps/baas-account-service/src/customers/customers.service.ts
	modified:   apps/core-bank-simulator/src/core-accounts/core-accounts.service.ts
	modified:   apps/core-bank-simulator/src/core-accounts/dto/create-core-account.dto.ts
	modified:   libs/baas-errors/src/baas.errors.ts
	modified:   libs/baas-errors/src/baas.exceptions.ts
	modified:   libs/baas-interfaces/src/accounts/account.interface.ts