[DebitCard] Refactor the Update Debit Card APIs
[apps] baas-debit-card-service

Refactoring the baas-debit-card-service APIs that activate, cancel, update
pin, and update limits by removing the separate resources I created in the
app and migrating the logic to the debit-cards.controller.ts. Added an
update method to the debit-cards.service.ts to call the core-bank-simulator
to perform the update. Refactored APIs are:
- PATCH /v1/debit-cards/:debitCardId/activate
- PATCH /v1/debit-cards/:debitCardId/cancel
- PATCH /v1/debit-cards/:debitCardId/pin
- PATCH /v1/debit-cards/:debitCardID/limits
-
Refactored the baas-debit-card-service to handle the new response
message format from the core-bank-simulator that encapsulates the
response in the "data" field, so the response looks like:
{
  "data": {
    "id": "xyz",
    "field_1": "value_1",
    ...
    "field_n": "value_n"
  }
}