[CoreBank] Refactored BaaS Services to Handle New Core Bank Response Format
Apps: baas-customer-service, baas-account-service, baas-debit-card-service

Updated all of the Rest APIs for the baas-customer-service, baas-account-service,
and baas-debit-card-service to handle the new response format sent from the
core-bank-simulator that encapsulates the message in the "data" field. Also,
refactored the unit tests to use the "setMockConfigService" function to set
the node environment required to run the jest unit tests instead of hard-coding
all the mockConfigService setup in each test file.