[BaasPlatform] Added RequestId to all Log Messages
Added a Request Id as a field in all JSON loge messages by adding a
RequestIdInterceptor that checks if the API request has the 'Request-Id'
field set in the header, if it does not then the interceptor generates
a unique request Id and stores it in node's AsyncLocalStorage. The
interceptor also adds the 'Request-Id' to response header, so that it
can be used to debug errors.
-
When writing a log message, the WinstonLoggerService retrieves the
'requestId' from AsyncLocalStorage and if it is set it adds it to the
log message.
-
Add the requestId logging to the baas-customer-service,
baas-account-service, baas-debit-card-service, and the
core-bank-simulator.